+++ src/js/app/loader/context.js
@@ -129,11 +129,12 @@ let Loader = function (store, api) {
             { ownErrorHandler: true },
             address,
             (json) => {
-                this.refresh()
                 this.notifyUser(<Message code='account.message.addressNew' message='Added new address' />, 'success')
 
                 let route = this.store.getState().app.route
                 app.getLoader('node').loadMaster(route.route, route.parameters)
+
+                app.getLoader('context').refresh()
             },
             (json) => {
                 this.notifyUser(<Message {...json} />, 'error')
@@ -148,11 +149,12 @@ let Loader = function (store, api) {
             { ownErrorHandler: true },
             address,
             (json) => {
-                this.refresh()
                 this.notifyUser(<Message code='account.message.addressUpdated' message='Updated address' />, 'success')
 
                 let route = this.store.getState().app.route
                 app.getLoader('node').loadMaster(route.route, route.parameters)
+
+                app.getLoader('context').refresh()
             },
             (json) => {
                 this.notifyUser(<Message {...json} />, 'error')
@@ -167,11 +169,12 @@ let Loader = function (store, api) {
             { ownErrorHandler: true },
             address,
             (json) => {
-                this.refresh()
                 this.notifyUser(<Message code='account.message.addressRemoved' message='Removed address' />, 'success')
 
                 let route = this.store.getState().app.route
                 app.getLoader('node').loadMaster(route.route, route.parameters)
+
+                app.getLoader('context').refresh()
             },
             (json) => {
                 this.notifyUser(<Message {...json} />, 'error')
@@ -190,6 +193,8 @@ let Loader = function (store, api) {
 
                 let route = this.store.getState().app.route
                 app.getLoader('node').loadMaster(route.route, route.parameters)
+
+                app.getLoader('context').refresh()
             },
             (json) => {
                 this.notifyUser(<Message {...json} />, 'error')
@@ -208,6 +213,8 @@ let Loader = function (store, api) {
 
                 let route = this.store.getState().app.route
                 app.getLoader('node').loadMaster(route.route, route.parameters)
+
+                app.getLoader('context').refresh()
             },
             (json) => {
                 this.notifyUser(<Message {...json} />, 'error')
@@ -222,7 +229,6 @@ let Loader = function (store, api) {
             { ownErrorHandler: true },
             user,
             (json) => {
-                this.refresh()
                 this.notifyUser(<Message code='account.message.update' message='Account data updated' />, 'success')
 
                 this.store.dispatch({
@@ -230,6 +236,8 @@ let Loader = function (store, api) {
                     data: json,
                     id: user.email,
                 })
+
+                app.getLoader('context').refresh()
             },
             (json) => {
                 this.notifyUser(<Message {...json} />, 'error')
@@ -279,7 +287,8 @@ let Loader = function (store, api) {
                 newPassword: newPassword,
             },
             (json) => {
-                this.refresh()
+                app.getLoader('context').refresh()
+
                 this.notifyUser(<Message code='account.message.passwordUpdate' message='Password updated' />, 'success')
             },
             (json) => {